requirements.txt
locust
websocket-client
dotenv
